+++ CHANGELOG.md
@@ -17,7 +17,8 @@
 
 #### Bug Fixes
 * `EMP-11603` Fixed an issue where the `onAirplayStatusChanged` callback fired multiple times for a single status update.
-* `EMP-11587` Resolved an issue during *Airplay* where incoming calls caused a sound from the playback to play over the phone's ringtone. 
+* `EMP-11587` Resolved an issue during *Airplay* where incoming calls caused a sound from the playback to play over the phone's ringtone.
+* `EMP-11599` Bringing up the iOS Control Center no longer pauses local playback.
 
 ## 2.0.86